Make WLDateTransformer a generic iso8601 transformer.
Before this change WLDateTransformer accepted CPDate initWithString style dates, with or without the ending +0000. This is not generally useful since most APIs use ISO8601 formatted dates.

This change makes WLDateTransformer a general ISO8601 transformer.

If the old behaviour is required, just create a custom transformer containing the code as it was prior to this commit.