Allow remote objects to not care about some deferred properties for loading purposes.
For instance perhaps there's an optional field the server might send, or might not. If it does not, we should not be issuing a new GET for every ensureLoaded.

With this fix an object aware of such a field on the server can simply override - deferredProperties and remove the field from the list.